LayoutTests/accessibility/custom-elements/orientation-expected.txt

This tests that aria fallback roles work correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS accessibilityController.accessibleElementById("group-1").orientation is "AXOrientation: AXVerticalOrientation"
PASS accessibilityController.accessibleElementById("group-2").orientation is "AXOrientation: AXHorizontalOrientation"
PASS successfullyParsed is true

TEST COMPLETE
